+++ src/const.ts
@@ -1,8 +1,10 @@
 export const ColorOK = 'rgba(55, 135, 45, 0.9)';
 export const ColorDisaster = 'rgba(196, 22, 42, 0.9)';
+export const ColorCritical = 'rgba(205, 49, 49, 0.9)';
 export const ColorHigh = 'rgba(217, 92, 29, 0.9)';
 export const ColorAverage = 'rgba(250, 100, 0, 0.99)';
 export const ColorWarning = 'rgba(224, 180, 0, 0.99)';
+export const ColorMinor = 'rgba(190, 213, 255, 0.99)';
 export const ColorInformation = 'rgba(138, 184, 255, 0.99)';
 export const ColorDisable = 'rgba(128, 128, 128, 0.9)';
 export const ColorNa = 'rgba(20, 60, 120, 0.9)';